+++ src/lastversion/__about__.py
@@ -1,4 +1,4 @@
 """Package metadata"""
 
-__version__ = "3.6.4"
+__version__ = "3.6.5"
 __self__ = "dvershinin/lastversion"